# jrs-hr-module/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JRS HR Notification and Profile API",
    "version": "0.1.0",
    "description": "The team owns sign-in; there is no password login endpoint in this module. The upstream authentication transport is pending agreement and is not specified by this document. An installed server adapter must verify it. GET /api/auth/me checks that identity and returns a token bound to its subject/session/HR mapping. Retain the separate jrs.hr.sid CSRF-session cookie and send x-csrf-token for mutations. A changed upstream identity requires refreshing /api/auth/me. The module cookie alone never authenticates a user. POST/PATCH/PUT/DELETE with a supplied Origin must match APP_ORIGIN. Log success means provider acceptance, not confirmed inbox delivery. PREVIEW/PENDING/FAILED are not successful history. No public workflow-send endpoint."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin API (including Vite proxy)"
    }
  ],
  "paths": {
...
    "/api/auth/csrf": {
      "get": {
        "operationId": "csrf",
        "tags": [
          "Session"
        ],
        "summary": "Create or retrieve anonymous/session CSRF token",
        "security": [],
        "x-team-identity-required": false,
        "description": "Public module endpoint; it does not grant HR access.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csrfToken": {
                      "type": "string",
                      "maxLength": 64,
                      "pattern": "^[a-f0-9]{64}$"
                    }
                  },
                  "required": [
                    "csrfToken"
                  ]
                },
                "examples": {
                  "token": {
                    "summary": "Illustrative token; use the live session token",
                    "value": {
                      "csrfToken": "0000000000000000000000000000000000000000000000000000000000000000"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "malformedJson": {
                    "summary": "Example malformed request",
                    "value": {
                      "error": {
                        "message": "Malformed JSON body."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Verified team sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "teamSignIn": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Continue through the team sign-in page."
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or HR authorization rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "csrf": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "CSRF token missing or expired. Refresh and retry."
                      }
                    }
                  },
                  "origin": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Untrusted request origin."
                      }
                    }
                  },
                  "inactiveHr": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "HR profile is no longer authorized."
                      }
                    }
                  },
                  "changedSession": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "The team session changed. Refresh before retrying."
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Missing or inaccessible record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "inaccessible": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Record not found or not accessible."
                      }
                    }
                  },
                  "missingFile": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "File no longer available."
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Unique name conflict, including archived templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "reservedName": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "A record with that name already exists (including archived templates)."
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 100 KB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "largeBody": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Request body is too large."
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidFields": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Please check the entered values.",
                        "fields": [
                          {
                            "field": "body",
                            "message": "Unsupported template variable"
                          }
                        ]
                      }
                    }
                  },
                  "upload": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Upload one image, no larger than 2 MB."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "failed": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Something went wrong. Please try again."
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Team identity service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "upstream": {
                    "summary": "Upstream adapter unavailable; no identity fallback",
                    "value": {
                      "error": {
                        "message": "Something went wrong. Please try again."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/config": {
      "get": {
        "operationId": "authConfig",
        "tags": [
          "Session"
        ],
        "summary": "Read available sign-in entry points",
        "security": [],
        "x-team-identity-required": false,
        "description": "Public module endpoint; it does not grant HR access.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loginUrl": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    },
                    "adapterConfigured": {
                      "type": "boolean"
                    },
                    "developmentLogin": {
                      "type": "boolean",
                      "description": "True only for an explicitly configured local standalone environment."
                    }
                  },
                  "required": [
                    "loginUrl",
                    "adapterConfigured",
                    "developmentLogin"
                  ]
                },
                "examples": {
                  "pending": {
                    "summary": "No sign-in adapter is available",
                    "value": {
                      "loginUrl": null,
                      "adapterConfigured": false,
                      "developmentLogin": false
                    }
                  },
                  "registered": {
                    "summary": "Illustrative team adapter registration",
                    "value": {
                      "loginUrl": "https://team.example.test/sign-in",
                      "adapterConfigured": true,
                      "developmentLogin": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "malformedJson": {
                    "summary": "Example malformed request",
                    "value": {
                      "error": {
                        "message": "Malformed JSON body."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Verified team sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "teamSignIn": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Continue through the team sign-in page."
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF, origin or HR authorization rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "csrf": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "CSRF token missing or expired. Refresh and retry."
                      }
                    }
                  },
                  "origin": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Untrusted request origin."
                      }
                    }
                  },
                  "inactiveHr": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "HR profile is no longer authorized."
                      }
                    }
                  },
                  "changedSession": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "The team session changed. Refresh before retrying."
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Missing or inaccessible record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "inaccessible": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Record not found or not accessible."
                      }
                    }
                  },
                  "missingFile": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "File no longer available."
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Unique name conflict, including archived templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "reservedName": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "A record with that name already exists (including archived templates)."
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 100 KB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "largeBody": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Request body is too large."
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalidFields": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Please check the entered values.",
                        "fields": [
                          {
                            "field": "body",
                            "message": "Unsupported template variable"
                          }
                        ]
                      }
                    }
                  },
                  "upload": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Upload one image, no larger than 2 MB."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "failed": {
                    "summary": "Illustrative example; IDs are not live values",
                    "value": {
                      "error": {
                        "message": "Something went wrong. Please try again."
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Team identity service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "upstream": {
                    "summary": "Upstream adapter unavailable; no identity fallback",
                    "value": {
                      "error": {
                        "message": "Something went wrong. Please try again."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/docs": {
      "get": {
        "operationId": "swagger",
        "tags": [
          "Documentation"
        ],
        "summary": "Open Swagger UI",
        "security": [],
        "x-team-identity-required": false,
        "description": "Public module endpoint; it does not grant HR access.",
        "responses": {
          "200": {
            "description": "Swagger HTML page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "301": {
            "description": "Redirect to trailing slash"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jrs.hr.sid",
        "description": "Module CSRF session only. A separately verified upstream identity is also mandatory; its cookie or token transport remains to be agreed."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "64 lowercase hex characters bound to the current session."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "field",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}